Conflicts:
	ReMotor.mcp

+++ Sources/led.h
@@ -20,8 +20,8 @@
 void led_init(void);
 
 void led(u8 brightness);
-void led_blink (u8 dim_on, u8 dim_off, rti_time t_on, rti_time t_off, rti_time dur, int end_dim);
-void led_fade (rti_time t_trans, rti_time dur, int end_dim);
+void led_blink (u8 dim_on, u8 dim_off, rti_time t_on, rti_time t_off, rti_time dur, s16 end_dim);
+void led_fade (rti_time t_trans, rti_time dur, s16 end_dim);
 void led_stop();
 
 #endif /* __LED_H__ */
